doing unit tests
